Refactor: Split JobZonePage into smaller components
Refactors JobZonePage.tsx into separate components for the header, input section, progress indicator, and resume display sections, improving code organization and maintainability.
